requirements.txt
streamlit>=1.22.0
openai>=1.0.0
python-dotenv>=1.0.0
langchain>=0.0.267
faiss-cpu>=1.7.4
tiktoken>=0.4.0
numpy>=1.24.0
fastapi>=0.95.0
uvicorn>=0.22.0

# Document processing libraries
PyPDF2>=3.0.0
python-docx>=0.8.11
pandas>=1.5.0
openpyxl>=3.1.0
beautifulsoup4>=4.11.0
pymupdf>=1.21.0
docx
fitz 
